balance factor

+++ 14-binary_tree_balance.c
@@ -6,7 +6,7 @@
  * Return: tree height
  */
 
-size_t binary_tree_height(const binary_tree_t *tree)
+int binary_tree_height(const binary_tree_t *tree)
 {
 	size_t left_height, right_height;
